requirements.txt
torch
torchvision
tqdm
numpy
Pillow
scipy
